// src/types/user.ts
import { Types } from "mongoose";

export interface UserType {
  id: number;
  name: string;
  username: string;
  email: string;
  address?: Address;
  phone: string;
  website: string;
  company?: Company;
  blogs:Types.ObjectId[];
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo?: GeoLocation;
}

export interface GeoLocation {
  lat: string;
  lng: string;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}